requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
django==4.0.3
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
django-split-settings==1.1.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.1.0
    # via -r requirements.in
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
